requirements/base.txt
amqp==5.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.3.0
cryptography==40.0.2
css-inline==0.10.3
decorator==5.1.1
Deprecated==1.2.13
Django==4.2.1
dj-rest-auth==7.0.1
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.26.0
django-admin-rangefilter==0.10.0
django-admin-sortable2==2.1.8
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-channels==0.7.0
django-colorfield==0.8.0
django-cors-headers==4.0.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-hashid-field==3.3.7
django-oauth-toolkit==2.2.0
django-redis==5.2.0
django-rest-passwordreset==1.3.0
django-timezone-field==5.0
django-upload-validator==1.1.6
django-user-accounts==3.2.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.5
ERPpeek==1.7.1
et-xmlfile==1.1.0
executing==1.2.0
factory-boy==3.3.0
Faker==25.9.1
h11==0.14.0
hashids==1.3.1
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.13.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msgpack==1.0.5
mysql-connector-python==8.0.33
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytest-xdist==3.8.0
pytest-django==4.8.0
python-slugify==8.0.1
pytz==2023.3
redis==4.5.5
paymongo-python==0.1.0
requests==2.30.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
sentry-sdk==2.22.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.15.0
xlwt==1.3.0
zope.interface==6.0
